chore: account pool (#176)
* fix git conflict

* fix: fix missing args

* test: list accounted pool test case

* initiate interfaces in margin keeper

* fix: add missing params

* fix: kv store issue in accounted pool module through margin call

* fix: pool address to be normal accaddress

* remove amm reference from accounted pool and updated hooks including the pool instance

* remove all cross dependency in accountedpool keeper

* chore: fix unit tests of margin invariant check

---------

+++ x/margin/keeper/check_max_open_positions_test.go
@@ -24,7 +24,7 @@ func TestCheckMaxOpenPositions_OpenPositionsBelowMax(t *testing.T) {
 
 	// Mock behavior
 	mockChecker.On("GetOpenMTPCount", ctx).Return(uint64(5))
-	mockChecker.On("GetMaxOpenPositions", ctx).Return(10)
+	mockChecker.On("GetMaxOpenPositions", ctx).Return(uint64(10))
 
 	err := k.CheckMaxOpenPositions(ctx)
 
@@ -46,7 +46,7 @@ func TestCheckMaxOpenPositions_OpenPositionsEqualToMax(t *testing.T) {
 
 	// Mock behavior
 	mockChecker.On("GetOpenMTPCount", ctx).Return(uint64(10))
-	mockChecker.On("GetMaxOpenPositions", ctx).Return(10)
+	mockChecker.On("GetMaxOpenPositions", ctx).Return(uint64(10))
 
 	err := k.CheckMaxOpenPositions(ctx)
 
@@ -68,7 +68,7 @@ func TestCheckMaxOpenPositions_OpenPositionsExceedMax(t *testing.T) {
 
 	// Mock behavior
 	mockChecker.On("GetOpenMTPCount", ctx).Return(uint64(11))
-	mockChecker.On("GetMaxOpenPositions", ctx).Return(10)
+	mockChecker.On("GetMaxOpenPositions", ctx).Return(uint64(10))
 
 	err := k.CheckMaxOpenPositions(ctx)
